+++ rust-peer/src/main.rs
@@ -1,10 +1,11 @@
+mod protocol;
+
 use anyhow::{Context, Result};
-use async_trait::async_trait;
 use clap::Parser;
 use futures::future::{select, Either};
-use futures::{io, AsyncRead, AsyncWrite, StreamExt};
-use libp2p::core::upgrade::{read_length_prefixed, write_length_prefixed};
-use libp2p::request_response::{self, ProtocolName, ProtocolSupport};
+use futures::StreamExt;
+
+use libp2p::request_response::{self, ProtocolSupport};
 use libp2p::{
     core::muxing::StreamMuxerBox,
     gossipsub, identify, identity,
@@ -21,6 +22,7 @@ use libp2p_quic as quic;
 use libp2p_webrtc as webrtc;
 use libp2p_webrtc::tokio::Certificate;
 use log::{debug, error, info, warn};
+use protocol::FileExchangeCodec;
 use std::iter;
 use std::net::{IpAddr, Ipv4Addr};
 use std::path::Path;
@@ -31,6 +33,8 @@ use std::{
     time::{Duration, Instant},
 };
 use tokio::fs;
+
+use crate::protocol::{FileExchangeProtocol, FileRequest};
 
 const TICK_INTERVAL: Duration = Duration::from_secs(15);
 const KADEMLIA_PROTOCOL_NAME: &[u8] = b"/universal-connectivity/lan/kad/1.0.0";
@@ -154,10 +158,12 @@ async fn main() -> Result<()> {
                         let file_id = String::from_utf8(message.data).unwrap();
                         info!("Received file {} from {:?}", file_id, message.source);
 
-                        let request_id = swarm
-                            .behaviour_mut()
-                            .request_response
-                            .send_request(&message.source.unwrap(), FileRequest(file_id.clone()));
+                        let request_id = swarm.behaviour_mut().request_response.send_request(
+                            &message.source.unwrap(),
+                            FileRequest {
+                                file_id: file_id.clone(),
+                            },
+                        );
                         info!(
                             "Requested file {} to {:?}: req_id:{:?}",
                             file_id, message.source, request_id
@@ -239,10 +245,9 @@ async fn main() -> Result<()> {
                         );
                     }
                     request_response::Message::Response { response, .. } => {
-                        let file_body = response.0;
                         info!(
                             "request_response::Message::Response: size:{}",
-                            file_body.len()
+                            response.file_body.len()
                         );
                         // TODO: store this file (in memory or disk) and provider it via Kademlia
                     }
@@ -426,90 +431,4 @@ async fn read_or_create_identity(path: &Path) -> Result<identity::Keypair> {
     info!("Generated new identity and wrote it to {}", path.display());
 
     Ok(identity)
-}
-
-// Simple file exchange protocol
-
-#[derive(Debug, Clone)]
-struct FileExchangeProtocol();
-#[derive(Clone)]
-struct FileExchangeCodec();
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct FileRequest(String);
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FileResponse(Vec<u8>);
-
-impl ProtocolName for FileExchangeProtocol {
-    fn protocol_name(&self) -> &[u8] {
-        "/universal-connectivity-file/1".as_bytes()
-    }
-}
-
-#[async_trait]
-impl request_response::Codec for FileExchangeCodec {
-    type Protocol = FileExchangeProtocol;
-    type Request = FileRequest;
-    type Response = FileResponse;
-
-    async fn read_request<T>(
-        &mut self,
-        _: &FileExchangeProtocol,
-        io: &mut T,
-    ) -> io::Result<Self::Request>
-    where
-        T: AsyncRead + Unpin + Send,
-    {
-        let vec = read_length_prefixed(io, 1_000_000).await?;
-
-        if vec.is_empty() {
-            return Err(io::ErrorKind::UnexpectedEof.into());
-        }
-
-        Ok(FileRequest(String::from_utf8(vec).unwrap()))
-    }
-
-    async fn read_response<T>(
-        &mut self,
-        _: &FileExchangeProtocol,
-        io: &mut T,
-    ) -> io::Result<Self::Response>
-    where
-        T: AsyncRead + Unpin + Send,
-    {
-        let vec = read_length_prefixed(io, 500_000_000).await?; // update transfer maximum
-
-        if vec.is_empty() {
-            return Err(io::ErrorKind::UnexpectedEof.into());
-        }
-
-        Ok(FileResponse(vec))
-    }
-
-    async fn write_request<T>(
-        &mut self,
-        _: &FileExchangeProtocol,
-        io: &mut T,
-        FileRequest(data): FileRequest,
-    ) -> io::Result<()>
-    where
-        T: AsyncWrite + Unpin + Send,
-    {
-        write_length_prefixed(io, data).await?;
-
-        Ok(())
-    }
-
-    async fn write_response<T>(
-        &mut self,
-        _: &FileExchangeProtocol,
-        io: &mut T,
-        FileResponse(data): FileResponse,
-    ) -> io::Result<()>
-    where
-        T: AsyncWrite + Unpin + Send,
-    {
-        write_length_prefixed(io, data).await?;
-
-        Ok(())
-    }
 }
